requirements.txt
Django~=5.2.4
python-decouple~=3.8
djangorestframework~=3.16.0
django-leaflet==0.32.0